+++ R/Class_tskrrHeterogeneous.R
@@ -4,20 +4,16 @@
 #' \code{\link[xnet:tskrr-class]{tskrr}} specifically for
 #' heterogeneous networks.
 #'
-#' @slot y the matrix with responses
-#' @slot k the eigen decomposition of the kernel matrix for the rows
+#' @slot y an adjacency matrix.
+#' @slot k an object of class \code{\link[=dataClasses]{gramData}}
 #' @slot lambda.k the lambda value used for k
 #' @slot pred the matrix with the predictions
-#' @slot g the eigen decomposition of the kernel matrix for the columns
+#' @slot g an object of class \code{\link[=dataClasses]{gramData}}
 #' @slot lambda.g the lambda value used for g
 #' @slot has.hat a logical value indicating whether the kernel hat matrices
 #' are stored in the object.
 #' @slot Hk the kernel hat matrix for the rows.
 #' @slot Hg the kernel hat matrix for the columns.
-#' @slot labels a list with elements \code{k} and \code{g} (see
-#' \code{\link{tskrr-class}}).
-#'  If any element is \code{NA}, the labels used
-#' are integers indicating the row resp column number.
 #'
 #' @include Class_tskrr.R
 #' @rdname tskrrHeterogeneous-class
@@ -26,13 +22,10 @@
 #' @exportClass tskrrHeterogeneous
 setClass("tskrrHeterogeneous",
          contains = "tskrr",
-         slots = c(g = "eigen",
+         slots = c(g = "gramData",
                    lambda.g = "numeric",
                    Hg = "matrix"),
          prototype = list(lambda.g = 1e-4,
-                          g = structure(list(vectors = matrix(0),
-                                             values = numeric(1)),
-                                        class = "eigen"),
                           Hg = matrix(0)
                           )
          )