+++ R/Class_tskrrHomogeneous.R
@@ -4,8 +4,8 @@
 #' \code{\link[xnet:tskrr-class]{tskrr}} specifically for
 #' homogeneous networks.
 #'
-#' @slot y the matrix with responses
-#' @slot k the eigen decomposition of the kernel matrix for the rows
+#' @slot y an adjacency matrix.
+#' @slot k an object of class \code{\link[=dataClasses]{gramData}}
 #' @slot lambda.k the lambda value used for k
 #' @slot pred the matrix with the predictions
 #' @slot symmetry a character value that can have the possible values
@@ -15,10 +15,6 @@
 #' @slot has.hat a logical value indicating whether the kernel hat matrices
 #' are stored in the object.
 #' @slot Hk the kernel hat matrix for the rows.
-#' @slot labels a list with elements \code{k} and \code{g} (see
-#' \code{\link{tskrr-class}}). For homogeneous networks, \code{g}
-#' is always \code{NA}. If \code{k} is \code{NA}, the labels used
-#' are integers indicating the row resp column number.
 #'
 #' @include Class_tskrr.R
 #' @rdname tskrrHomogeneous-class