+++ NEWS.md
@@ -7,6 +7,10 @@ Currently in development as 0.1.12
 * `alpha()` is now called `get_alpha()` to avoid conflicts with other packages. The old function is still exported, but now gives a warning. It will be removed in a next version.
 * The function `get_kernel` (deprecated in 0.1.10) is now completely removed from the package.
 
+The `tskrr` classes have been rewritten to contain the `gramData` classes. This also removed the slot `labels` from the classes. Care is taken to keep old code compatible, but any code that relied directly on the structure of these classes will break. Following changes might break your code as well:
+
+* `valid_dimensions` now tests objects of `gramData` classes.
+
 ### New features
 
 * `tskrr` is now a S4 generic to allow for other input types.